Fixed shopping cart icon number
