+++ src/Components/NavBar.tsx
@@ -5,7 +5,7 @@ import { CSSProperties, makeStyles } from "@material-ui/styles";
 import { routes } from "../Navigation/routes";
 import { Link } from "react-router-dom";
 import DropDown from "./DropDown";
-import Cart from "./Cart";
+import CartButton from "./CartButton";
 
 const useStyles = makeStyles({
   root: {
@@ -36,7 +36,7 @@ function NavBar() {
   };
 
   return (
-    <AppBar position="static" className={classes.root}>
+    <AppBar position='static' className={classes.root}>
       <Toolbar>
         <IconButton
           component={Link}
@@ -47,7 +47,7 @@ function NavBar() {
           }}
           onClick={() => handleActivePage("home")}
         >
-          <img src={Logo} style={logoStyle} alt="logo" />
+          <img src={Logo} style={logoStyle} alt='logo' />
         </IconButton>
         <div style={wrapperDiv}>
           <Hidden smDown>
@@ -101,7 +101,7 @@ function NavBar() {
                   }}
                   onClick={() => handleActivePage(name)}
                 >
-                  <Cart />
+                  <CartButton />
                 </IconButton>
               ))}
             </div>
